General work

+++ remote/gitpoll.sh
@@ -1,4 +1,6 @@
 #!/bin/bash -x
+export PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/home/easydeploy/bin:/root/bin
+
 cd /home/easydeploy/deployment
 export EASYDEPLOY_HOST_IP=$(/sbin/ifconfig eth0 | sed -En 's/127.0.0.1//;s/.*inet (addr:)?(([0-9]*\.){3}[0-9]*).*/\2/p')
 
@@ -7,12 +9,12 @@ export EASYDEPLOY_HOST_IP=$(/sbin/ifconfig eth0 | sed -En 's/127.0.0.1//;s/.*ine
 while true
 do
     su  easydeploy -c "git fetch &> .build_log.txt"
-    if [ -s .build_log.txt ]
+    if [ -s .build_log.txt ] && [ ! -f /tmp/.install-in-progress ]
     then
-       su easydeploy -c "git revert Dockerfile; git pull"
+       su easydeploy -c "rm Dockerfile; git pull"
        if su easydeploy -c "/home/easydeploy/bin/build.sh 2>&1 | tee /tmp/build.out"
        then
-           supervisorctl restart $(cat /var/easydeploy/share/.config/component):              echo "Component restarted"
+           /home/easydeploy/bin/restart-component.sh && echo "Component restarted"
        else
             /home/easydeploy/bin/notify.sh "Build of $(cat /var/easydeploy/share/.config/component) failed on ${EASYDEPLOY_HOST_IP} " < /tmp/build.out
             echo "Docker build failed, no redeploy attempted."
